refactor(reservation): 유효성 검사 로직을 reservation 도메인으로 이동
